refactor(shared): de-dup SVG sanitiser + S3 branding utils, doc shared layer (#480 tasks 2-4)
Completes issue #480 on top of the finance extraction:

- SVG sanitiser: single impl at @givernance/shared/lib/svg-sanitiser (subpath);
  packages/api + packages/worker lib/svg-sanitiser.ts become thin re-exports.
- S3 branding: branding helpers (putBrandingObject, getBrandingObject,
  deleteBrandingPrefix, brandingPublicUrl) moved to @givernance/shared/lib/s3-branding
  with S3Client injection; both lib/s3.ts pre-bind their own client + bucket env and
  delegate (call sites unchanged). Generic S3 wiring / lib-storage streaming / SSE stay
  per-package.
- docs/02-reference-architecture.md: new "Shared business layer" section documenting the
  DB/S3-client dependency-injection pattern + the subpath-export convention (heavy/DB
  modules kept out of the root barrel per ADR-013).

Subpath exports only — root barrel untouched. typecheck + biome + build green.
