+++ packages/api/src/lib/s3.ts
@@ -4,10 +4,15 @@ import type { Readable } from "node:stream";
 import {
   DeleteObjectsCommand,
   GetObjectCommand,
-  ListObjectsV2Command,
   PutObjectCommand,
   S3Client,
 } from "@aws-sdk/client-s3";
+import {
+  brandingPublicUrl as brandingPublicUrlShared,
+  deleteBrandingPrefix as deleteBrandingPrefixShared,
+  getBrandingObject as getBrandingObjectShared,
+  putBrandingObject as putBrandingObjectShared,
+} from "@givernance/shared/lib/s3-branding";
 import { env } from "../env.js";
 
 const s3 = new S3Client({
@@ -98,42 +103,19 @@ export async function fetchCampaignObject(s3Path: string): Promise<{
 
 // ─── Branding bucket helpers (Epic #286) ───────────────────────────────────
 //
-// The branding bucket is **public-read** at the bucket level so donor-
-// facing surfaces (Keycloak login template, donation page) can hit
-// object URLs directly with `Cache-Control: public, max-age=31536000,
-// immutable`. The keys themselves are content-addressed
-// (`{org_id}/logo/{logo_id}/{variant}.{ext}`) so a stale CDN edge can
-// never leak a logo across logo versions.
-//
-// IMPORTANT — defense-in-depth:
-//   - We do NOT set `ACL: private` on these uploads (would override the
-//     bucket-level policy and 403 the donor).
-//   - We do NOT set SSE — encryption-at-rest is enabled at the bucket
-//     level (Scaleway default + MinIO via `MINIO_KMS_SECRET_KEY`); a
-//     per-object override would force a key dance for a value that
-//     is, by design, public.
+// The branding-asset implementations live in the shared package
+// (`@givernance/shared/lib/s3-branding`, issue #480) so the API and the
+// worker share ONE copy. The wrappers below pre-bind this package's own
+// `s3` client + `env` bucket/URL config so every existing call site keeps
+// its original signature unchanged. See the shared module for the full
+// public-read / content-addressed-key / no-per-object-ACL rationale.
 
 /**
  * Upload a branding asset (original or derived variant) to the public-
- * read branding bucket. The Cache-Control + Content-Type are baked in
- * here so every upload site stays consistent.
+ * read branding bucket. Returns the key.
  */
-export async function putBrandingObject(
-  key: string,
-  body: Buffer,
-  contentType: string,
-): Promise<string> {
-  await s3.send(
-    new PutObjectCommand({
-      Bucket: env.S3_BRANDING_BUCKET,
-      Key: key,
-      Body: body,
-      ContentType: contentType,
-      // Long-lived public cache — content-addressed keys make this safe.
-      CacheControl: "public, max-age=31536000, immutable",
-    }),
-  );
-  return key;
+export function putBrandingObject(key: string, body: Buffer, contentType: string): Promise<string> {
+  return putBrandingObjectShared(s3, env.S3_BRANDING_BUCKET, key, body, contentType);
 }
 
 /**
@@ -141,71 +123,31 @@ export async function putBrandingObject(
  * handler to embed the active logo in the on-the-fly PDF preview).
  * Returns `null` on 404 so callers can degrade gracefully.
  */
-export async function getBrandingObject(key: string): Promise<Buffer | null> {
-  try {
-    const out = await s3.send(
-      new GetObjectCommand({
-        Bucket: env.S3_BRANDING_BUCKET,
-        Key: key,
-      }),
-    );
-    const body = out.Body as Readable | undefined;
-    if (!body) return null;
-    const chunks: Buffer[] = [];
-    for await (const chunk of body) {
-      chunks.push(chunk as Buffer);
-    }
-    return Buffer.concat(chunks);
-  } catch (err) {
-    const e = err as { name?: string; $metadata?: { httpStatusCode?: number } };
-    if (e?.name === "NoSuchKey" || e?.$metadata?.httpStatusCode === 404) {
-      return null;
-    }
-    throw err;
-  }
+export function getBrandingObject(key: string): Promise<Buffer | null> {
+  return getBrandingObjectShared(s3, env.S3_BRANDING_BUCKET, key);
 }
 
 /**
  * Delete every object under a prefix. Used by the `branding.gc_asset`
  * worker job to clean up after a soft-deleted asset (and by tenant
- * offboarding to wipe `{org_id}/`). Best-effort — we paginate the
- * `ListObjectsV2` cursor and batch deletes in chunks of 1000 (S3 cap).
+ * offboarding to wipe `{org_id}/`).
  */
-export async function deleteBrandingPrefix(prefix: string): Promise<number> {
-  let totalDeleted = 0;
-  let continuationToken: string | undefined;
-  do {
-    const list = await s3.send(
-      new ListObjectsV2Command({
-        Bucket: env.S3_BRANDING_BUCKET,
-        Prefix: prefix,
-        ContinuationToken: continuationToken,
-      }),
-    );
-    const objects = (list.Contents ?? []).map((obj) => ({ Key: obj.Key as string }));
-    if (objects.length > 0) {
-      await s3.send(
-        new DeleteObjectsCommand({
-          Bucket: env.S3_BRANDING_BUCKET,
-          Delete: { Objects: objects, Quiet: true },
-        }),
-      );
-      totalDeleted += objects.length;
-    }
-    continuationToken = list.IsTruncated ? list.NextContinuationToken : undefined;
-  } while (continuationToken);
-  return totalDeleted;
+export function deleteBrandingPrefix(prefix: string): Promise<number> {
+  return deleteBrandingPrefixShared(s3, env.S3_BRANDING_BUCKET, prefix);
 }
 
 /**
  * Compose the public URL of a branding object. Defaults to
- * `${S3_ENDPOINT}/${S3_BRANDING_BUCKET}/${key}` so local dev (MinIO)
- * works without configuration. Production overrides via
+ * `${S3_ENDPOINT}/${S3_BRANDING_BUCKET}/${key}` so local dev works
+ * without configuration. Production overrides via
  * `KEYCLOAK_LOGO_PUBLIC_URL_BASE` to point at the CDN.
  */
 export function brandingPublicUrl(key: string): string {
-  const base = env.KEYCLOAK_LOGO_PUBLIC_URL_BASE ?? `${env.S3_ENDPOINT}/${env.S3_BRANDING_BUCKET}`;
-  return `${base.replace(/\/+$/, "")}/${key}`;
+  return brandingPublicUrlShared(key, {
+    endpoint: env.S3_ENDPOINT,
+    brandingBucket: env.S3_BRANDING_BUCKET,
+    publicUrlBase: env.KEYCLOAK_LOGO_PUBLIC_URL_BASE,
+  });
 }
 
 // ─── Bulk Import bucket helpers (Epic #373) ───────────────────────────
