+++ packages/shared/package.json
@@ -12,6 +12,8 @@
     "./validators": "./src/validators/index.ts",
     "./constants": "./src/constants/index.ts",
     "./finance/reporting": "./src/finance/reporting/index.ts",
+    "./lib/svg-sanitiser": "./src/lib/svg-sanitiser.ts",
+    "./lib/s3-branding": "./src/lib/s3-branding.ts",
     "./i18n": "./src/i18n/locales.ts",
     "./contracts/branding": "./src/contracts/branding.ts",
     "./postal-export-mode": "./src/postal-export-mode.ts",
@@ -25,9 +27,13 @@
   },
   "dependencies": {
     "drizzle-orm": "^0.45.2",
-    "@sinclair/typebox": "^0.34.12"
+    "@aws-sdk/client-s3": "^3.1057.0",
+    "@sinclair/typebox": "^0.34.12",
+    "dompurify": "^3.4.7",
+    "jsdom": "^29.1.1"
   },
   "devDependencies": {
+    "@types/jsdom": "^28.0.3",
     "@types/pg": "^8.20.0",
     "typescript": "^6.0.3"
   }
